moved ExportModal.tsx to shared
