+++ ui/components/designs/patterns/MesheryPatterns.tsx
@@ -68,7 +68,7 @@ import { DefaultTableCell, SortableTableCell } from '../connections/common';
 import DefaultError from '../General/error-404/index';
 import CAN from '@/utils/can';
 import { keys } from '@/utils/permission_constants';
-import ExportModal from '../../ExportModal';
+import ExportModal from '../shared/Modal/ExportModal';
 import { useModal, Modal as SistentModal, ModalBody } from '@sistent/sistent';
 import PatternIcon from '@/assets/icons/Pattern';
 import DryRunIcon from '@/assets/icons/DryRunIcon';
